added exit states, multi-file operations, compression
Added multiple exit states for the API that reflect attempting to interact with non-existent devices.

Modified API to accept comma deiminated lists of filenames for 'delete_device' and 'download_device', allowing for commands that affect multiple files.

Modified web and API to download files to users as compressed ZIP files instead of PDF.